modules/SceAppMgrUser,SceDisplayUser & SceRtc/SceRtcUser: move all function inside kernel modules.
