+++ vita3k/modules/SceDisplay/SceDisplay.h
@@ -19,6 +19,38 @@
 
 #include <module/module.h>
 
+enum SceDisplaySetBufSync {
+    SCE_DISPLAY_SETBUF_IMMEDIATE = 0,
+    SCE_DISPLAY_SETBUF_NEXTFRAME = 1
+};
+
+enum SceDisplayErrorCode {
+    SCE_DISPLAY_ERROR_OK = 0,
+    SCE_DISPLAY_ERROR_INVALID_HEAD = 0x80290000,
+    SCE_DISPLAY_ERROR_INVALID_VALUE = 0x80290001,
+    SCE_DISPLAY_ERROR_INVALID_ADDR = 0x80290002,
+    SCE_DISPLAY_ERROR_INVALID_PIXELFORMAT = 0x80290003,
+    SCE_DISPLAY_ERROR_INVALID_PITCH = 0x80290004,
+    SCE_DISPLAY_ERROR_INVALID_RESOLUTION = 0x80290005,
+    SCE_DISPLAY_ERROR_INVALID_UPDATETIMING = 0x80290006,
+    SCE_DISPLAY_ERROR_NO_FRAME_BUFFER = 0x80290007,
+    SCE_DISPLAY_ERROR_NO_PIXEL_DATA = 0x80290008,
+    SCE_DISPLAY_ERROR_NO_OUTPUT_SIGNAL = 0x80290009
+};
+
+struct SceDisplayFrameBuf {
+    uint32_t size = 0;
+    Ptr<const void> base;
+    uint32_t pitch = 0;
+    uint32_t pixelformat = SCE_DISPLAY_PIXELFORMAT_A8B8G8R8;
+    uint32_t width = 0;
+    uint32_t height = 0;
+};
+
+EXPORT(SceInt32, _sceDisplaySetFrameBuf, const SceDisplayFrameBuf *pFrameBuf, SceDisplaySetBufSync sync);
+
+SceInt32 _sceDisplaySetFrameBuf(const SceDisplayFrameBuf *pParam, SceDisplaySetBufSync sync);
+
 BRIDGE_DECL(_sceDisplayGetFrameBuf)
 BRIDGE_DECL(_sceDisplayGetFrameBufInternal)
 BRIDGE_DECL(_sceDisplayGetMaximumFrameBufResolution)
