+++ vita3k/modules/SceDisplay/SceDisplayUser.cpp
@@ -17,34 +17,6 @@
 
 #include "SceDisplayUser.h"
 
-enum SceDisplaySetBufSync {
-    SCE_DISPLAY_SETBUF_IMMEDIATE = 0,
-    SCE_DISPLAY_SETBUF_NEXTFRAME = 1
-};
-
-enum SceDisplayErrorCode {
-    SCE_DISPLAY_ERROR_OK = 0,
-    SCE_DISPLAY_ERROR_INVALID_HEAD = 0x80290000,
-    SCE_DISPLAY_ERROR_INVALID_VALUE = 0x80290001,
-    SCE_DISPLAY_ERROR_INVALID_ADDR = 0x80290002,
-    SCE_DISPLAY_ERROR_INVALID_PIXELFORMAT = 0x80290003,
-    SCE_DISPLAY_ERROR_INVALID_PITCH = 0x80290004,
-    SCE_DISPLAY_ERROR_INVALID_RESOLUTION = 0x80290005,
-    SCE_DISPLAY_ERROR_INVALID_UPDATETIMING = 0x80290006,
-    SCE_DISPLAY_ERROR_NO_FRAME_BUFFER = 0x80290007,
-    SCE_DISPLAY_ERROR_NO_PIXEL_DATA = 0x80290008,
-    SCE_DISPLAY_ERROR_NO_OUTPUT_SIGNAL = 0x80290009
-};
-
-struct SceDisplayFrameBuf {
-    uint32_t size = 0;
-    Ptr<const void> base;
-    uint32_t pitch = 0;
-    uint32_t pixelformat = SCE_DISPLAY_PIXELFORMAT_A8B8G8R8;
-    uint32_t width = 0;
-    uint32_t height = 0;
-};
-
 EXPORT(int, sceDisplayGetFrameBuf) {
     return UNIMPLEMENTED();
 }
@@ -61,35 +33,8 @@ EXPORT(int, sceDisplayGetResolutionInfoInternal) {
     return UNIMPLEMENTED();
 }
 
-EXPORT(int, sceDisplaySetFrameBuf, const SceDisplayFrameBuf *pParam, SceDisplaySetBufSync sync) {
-    if (pParam == nullptr)
-        return SCE_DISPLAY_ERROR_OK;
-    if (pParam->size != sizeof(SceDisplayFrameBuf)) {
-        return RET_ERROR(SCE_DISPLAY_ERROR_INVALID_VALUE);
-    }
-    if (!pParam->base) {
-        return RET_ERROR(SCE_DISPLAY_ERROR_INVALID_ADDR);
-    }
-    if (pParam->pitch < pParam->width) {
-        return RET_ERROR(SCE_DISPLAY_ERROR_INVALID_PITCH);
-    }
-    if (pParam->pixelformat != SCE_DISPLAY_PIXELFORMAT_A8B8G8R8) {
-        return RET_ERROR(SCE_DISPLAY_ERROR_INVALID_PIXELFORMAT);
-    }
-    if (sync != SCE_DISPLAY_SETBUF_NEXTFRAME && sync != SCE_DISPLAY_SETBUF_IMMEDIATE) {
-        return RET_ERROR(SCE_DISPLAY_ERROR_INVALID_UPDATETIMING);
-    }
-
-    host.display.base = pParam->base;
-    host.display.pitch = pParam->pitch;
-    host.display.pixelformat = pParam->pixelformat;
-    host.display.image_size.x = pParam->width;
-    host.display.image_size.y = pParam->height;
-    ++host.frame_count;
-
-    MicroProfileFlip(nullptr);
-
-    return SCE_DISPLAY_ERROR_OK;
+EXPORT(SceInt32, sceDisplaySetFrameBuf, const SceDisplayFrameBuf *pFrameBuf, SceDisplaySetBufSync sync) {
+    return CALL_EXPORT(_sceDisplaySetFrameBuf, pFrameBuf, sync);
 }
 
 EXPORT(int, sceDisplaySetFrameBufForCompat) {
