+++ vita3k/modules/SceLibKernel/SceLibKernel.cpp
@@ -19,6 +19,8 @@
 #include <modules/module_parent.h>
 #include <v3kprintf.h>
 
+#include <../SceKernelThreadMgr/SceThreadmgr.h>
+
 #include <cpu/functions.h>
 #include <dlmalloc.h>
 #include <host/functions.h>
@@ -1547,28 +1549,8 @@ EXPORT(int, sceKernelWaitSemaCB, SceUID semaid, int signal, SceUInt *timeout) {
     return semaphore_wait(host.kernel, export_name, thread_id, semaid, signal, timeout);
 }
 
-// TODO figure out more about this struct
-struct SceKernelWaitSignalResult {
-    Address tls_address;
-    uint32_t dret;
-};
-
-// TODO figure out more about this struct
-struct SceKernelWaitSignalParams {
-    uint32_t reserved[2];
-    Ptr<SceKernelWaitSignalResult> result_ptr;
-};
-
 EXPORT(int, sceKernelWaitSignal, uint32_t unknown, uint32_t delay, uint32_t timeout, SceKernelWaitSignalParams *params) {
-    STUBBED("sceKernelWaitSignal");
-    const auto thread = lock_and_find(thread_id, host.kernel.threads, host.kernel.mutex);
-    LOG_TRACE("thread {} is waiting to get signaled", thread_id);
-    thread->signal.wait();
-    LOG_TRACE("thread {} gets signaled", thread_id);
-    if (params != nullptr) {
-        params->result_ptr.get(host.mem)->dret = 0;
-    }
-    return SCE_KERNEL_OK;
+    return CALL_EXPORT(_sceKernelWaitSignal, unknown, delay, timeout, params);
 }
 
 EXPORT(int, sceKernelWaitSignalCB) {
